secures and refactors resolvers
