+++ components/MCQ.tsx
@@ -20,12 +20,13 @@ import GameEndedView from './GameEndedView'
 import { useRouter } from 'next/navigation'
 
 type Props = {
-  gameId: string
+  gameId: string,
+  userId: string
 }
 
-const MCQ = ({ gameId }: Props) => {
+const MCQ = ({ gameId, userId }: Props) => {
   const router = useRouter();
-  const { userRole, userId } = useUserContext();
+  const { userRole } = useUserContext();
   const { game, loading, error, closeGame, finishGame, updateGameQuestion, addSpectatorToGame } = useGames({ gameId, userRole });
   const [selectedChoice, setSelectedChoice] = React.useState<number>(0);
   const [correctAnswers, setCorrectAnswers] = React.useState<number>(0);
@@ -35,11 +36,11 @@ const MCQ = ({ gameId }: Props) => {
 
   React.useEffect(() => {
     // if the user-Player tries to open a game that he's not a player of
-    if (userRole === Role.PLAYER && game.playerId !== userId) {
+    if (userRole === Role.PLAYER && game.playerId !== userId && !loading && !error) {
       router.push('/');
     }
 
-    if (game.gameType === GameType.mcq) {
+    if (game.gameType === GameType.open_ended && !loading && !error) {
       router.push(`/play/open-ended/${gameId}`)
     }
 
@@ -111,16 +112,18 @@ const MCQ = ({ gameId }: Props) => {
 
   React.useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === '1') {
-        setSelectedChoice(0);
-      } else if (event.key === '2') {
-        setSelectedChoice(1);
-      } else if (event.key === '3') {
-        setSelectedChoice(2);
-      } else if (event.key === '4') {
-        setSelectedChoice(3);
-      } else if (event.key === 'Enter') {
-        handleNext();
+      if (userRole === Role.PLAYER) {
+        if (event.key === '1') {
+          setSelectedChoice(0);
+        } else if (event.key === '2') {
+          setSelectedChoice(1);
+        } else if (event.key === '3') {
+          setSelectedChoice(2);
+        } else if (event.key === '4') {
+          setSelectedChoice(3);
+        } else if (event.key === 'Enter') {
+          handleNext();
+        }
       }
     }
 
@@ -181,7 +184,7 @@ const MCQ = ({ gameId }: Props) => {
         <CardHeader className='flex flex-row -items-center'>
           <CardTitle className="mr-5 text-center divide-y divide-zinc-600/50">
             <div>{game.currentQuestionIndex + 1}</div>
-            <div className="text-base text-slate-400">{game.questions.length}</div>
+            <div className="text-base text-slate-400">{game.totalQuestionsCount}</div>
           </CardTitle>
           <CardDescription className="flex-grow text-lg">
             {currentQuestion.question}
